File sending and read receipts are on board
File sending is a WIP, but most of the code is there now. I've
removed rt and st hashes and replaced them with a single transfer
hash where the key for each transfer is the file-id. Avatars also
use the transfer hash and will show up in the on-going transfer
list.

I've also added code where if a friend sends a message, but the
specific chat window isn't already focused, there'll be a little
asterisk in the frame title so we can tell we have unread messages.
The asterisk goes away when the window gains focus.