Bootstrap to several nodes and retry connections
When first starting up, bootstrap from 4 nodes at a time and then wait 5
seconds before trying them all again. This will ensure that we get
connected eventually to the network.

Also, changes for integer values to be enum symbols. JSON does not
accept symbols, so when saving to file, the symbol is converted to a
string.